Create DashGoProject.py
This is the first version of the DashGo Project. It contains one Breakdown for now called the "Revenue and Payable Breakdown" which shows The Track Title, Label Name, Distribution rate (What artist AND label receives), and Deal Share (What the artist receives). This code shows a better summary than what is shown in the DashGo csv file itself.